+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- add support for fetching available service types for managed database service
+
 ### Fixed
 - drop IP address family requirement from firewall rule