+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [4.6.0]
+
 ### Added
 - add support for fetching available service types for managed database service
 
@@ -164,7 +166,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.5.2...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.6.0...HEAD
+[4.6.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.5.2...v4.6.0
 [4.5.2]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.5.1...v4.5.2
 [4.5.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.5.0...v4.5.1
 [4.5.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.4.2...v4.5.0